Add Navbar, HeroSection and Explorer in Header
As the header section is growing so it is now in @/components/Header/
folder along with its individual components. Navbar got its own folder
as well.

Navbar currently holds a temporary logo and search bar on the lhs and on
rhs there is 'connect wallet' button and the hamburger menu which, for
the moment, contains some option text only. Below the navbar is hero
section which has a cycle of 5 images and some text. Then there is
explorer bar which shows different categories of NFT's to browse from.